0000834: Increased reports of vehicles disappearing
git-svn-id: http://svn2.xp-dev.com/svn/Fenix1053-MTARP@1363 64450a49-1f69-0410-a0a2-f5ebb52c4f5b

+++ src_trunk/resources/carshop-system/s_carshop_system.lua
@@ -63,6 +63,7 @@ end
 addEvent("buyCar", true)
 addEventHandler("buyCar", getRootElement(), buyCar)
 
+local destroyTimers = { }
 function makeCar(thePlayer, id, cost, col1, col2, x, y, z, rz, px, py, pz, prz)
 	local rx = 0
 	local ry = 0
@@ -115,20 +116,70 @@ function makeCar(thePlayer, id, cost, col1, col2, x, y, z, rz, px, py, pz, prz)
 		exports.global:givePlayerAchievement(thePlayer, 17) -- my ride
 		
 		setElementData(veh, "requires.vehpos", 1, false)
-		setTimer(checkVehpos, 3600000, 1, veh)
+		local timer = setTimer(checkVehpos, 3600000, 1, veh, insertid)
+		table.insert(destroyTimers, {timer, insertid})
 	end
 end
 
-function checkVehpos(veh)
+function checkVehpos(veh, insertid)
 	local requires = getElementData(veh, "requires.vehpos")
 	
 	if (requires) then
 		if (requires==1) then
 			local id = tonumber(getElementData(veh, "dbid"))
-			exports.irc:sendMessage("Removing vehicle #" .. id .. " (Did not get Vehpossed).")
-			destroyElement(veh)
-			local query = mysql_query(handler, "DELETE FROM vehicles WHERE id='" .. id .. "' LIMIT 1")
-			mysql_free_result(query)
+			
+			if (id==insertid) then
+				exports.irc:sendMessage("Removing vehicle #" .. id .. " (Did not get Vehpossed).")
+				destroyElement(veh)
+				local query = mysql_query(handler, "DELETE FROM vehicles WHERE id='" .. id .. "' LIMIT 1")
+				mysql_free_result(query)
+			end
 		end
 	end
 end
+
+-- VEHPOS
+function setVehiclePosition(thePlayer, commandName)
+	local veh = getPedOccupiedVehicle(thePlayer)
+	if not veh or getElementData(thePlayer, "realinvehicle") == 0 then
+		outputChatBox("You are not in a vehicle.", thePlayer, 255, 0, 0)
+	else
+		local playerid = getElementData(thePlayer, "dbid")
+		local owner = getElementData(veh, "owner")
+		local dbid = getElementData(veh, "dbid")
+		local TowingReturn = call(getResourceFromName("tow-system"), "CanTowTruckDriverVehPos", thePlayer) -- 2 == in towing and in col shape, 1 == colshape only, 0 == not in col shape
+		if (exports.global:isPlayerAdmin(thePlayer)) or (owner==playerid and TowingReturn == 0) or (exports.global:doesPlayerHaveItem(thePlayer, 3, dbid)) or (TowingReturn == 2) then
+			if (dbid<0) then
+				outputChatBox("This vehicle is not permanently spawned.", thePlayer, 255, 0, 0)
+			else
+				if (call(getResourceFromName("tow-system"), "CanTowTruckDriverGetPaid", thePlayer)) then
+					call(getResourceFromName("faction-system"), "addToFactionMoney", 24, 75)
+					call(getResourceFromName("faction-system"), "addToFactionMoney", 1, 75)
+				end
+				removeElementData(veh, "requires.vehpos")
+				local x, y, z = getElementPosition(veh)
+				local rx, ry, rz = getVehicleRotation(veh)
+				
+				local interior = getElementInterior(thePlayer)
+				local dimension = getElementDimension(thePlayer)
+				
+				local query = mysql_query(handler, "UPDATE vehicles SET x='" .. x .. "', y='" .. y .."', z='" .. z .. "', rotx='" .. rx .. "', roty='" .. ry .. "', rotz='" .. rz .. "', currx='" .. x .. "', curry='" .. y .. "', currz='" .. z .. "', currrx='" .. rx .. "', currry='" .. ry .. "', currrz='" .. rz .. "', interior='" .. interior .. "', currinterior='" .. interior .. "', dimension='" .. dimension .. "', currdimension='" .. dimension .. "' WHERE id='" .. dbid .. "'")
+				mysql_free_result(query)
+				setVehicleRespawnPosition(veh, x, y, z, rx, ry, rz)
+				outputChatBox("Vehicle spawn position set.", thePlayer)
+				
+				for key, value in ipairs(destroyTimers) do
+					if (tonumber(destroyTimers[key][2]) == dbid) then
+						local timer = destroyTimers[key][1]
+						
+						if (isTimer(timer)) then
+							killTimer(timer)
+							table.remove(destroyTimers, key)
+						end
+					end
+				end
+			end
+		end
+	end
+end
+addCommandHandler("vehpos", setVehiclePosition, false, false)
